Refactor TracrCorrespondence and fix bugs in it
